+++ client/src/app/app.component.ts
@@ -1,10 +1,19 @@
-import { Component } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {AppService} from './app.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-export class AppComponent {
-  title = 'app';
+export class AppComponent implements OnInit {
+  title = 'App!!';
+  msg: string;
+
+  constructor(private appService: AppService) {
+  }
+
+  ngOnInit() {
+    this.appService.testRoute().subscribe(data => this.msg = data['msg']);
+  }
 }