+++ postgresql/init.sql
@@ -1,0 +1,1 @@
+CREATE DATABASE users_dev;