+++ server/tests/__init__.py
@@ -1,0 +1,2 @@
+"""Config for running unittest"""
+config_name = 'testing'